requirements.txt
click==7.1.2
fastapi==0.68.0
h11==0.13.0
motor==3.0.0
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
starlette==0.14.2
typing_extensions==4.3.0
uvicorn==0.12.2
